Add texture support and basic resource management
- Added image loading with stb_image
- Added Texture bindable to manage texture data and resource view
- Changed vertex definition to support texture coordinates
- Updated vertexshader and pixelshader to support texture sampling
- Added ResourceManager class to manage resource loading and lifetime
- Scene loading is now done by the resource manager with paths to the
  required resources specified in scene.txt
- Added Material class to bundle shaders and material properties
- Moved matrix handling from cube into new Transform class
- Added MeshPrimitives to define vertex/index data for basic shapes
- Cube no longer directly handles any resources and can be generalised
  as a renderable object

+++ Graphics-Demo/src/d3d/indexbuffer.cpp
@@ -22,7 +22,7 @@ namespace d3d
 		DB_LOG("Creating IndexBuffer, IndexCount: " << m_indexCount << ", ByteWidth: " << m_byteWidth << '\n');
 
 		D3D11_BUFFER_DESC bufferDesc{};
-		bufferDesc.ByteWidth = m_byteWidth;
+		bufferDesc.ByteWidth = static_cast<UINT>(m_byteWidth);
 		bufferDesc.StructureByteStride = 0u;
 		bufferDesc.Usage = D3D11_USAGE_IMMUTABLE;
 		bufferDesc.BindFlags = D3D11_BIND_INDEX_BUFFER;
